Enable resizing and set fixed aspect ratio
Usually with resizing enabled you'd have to set a resize handler to
adjust the projection matrix to the new aspect ratio. However, if you
keep the same aspect ratio, you won't have to do that.

This commit calls glfwSetWindowAspectRatio in RootTask::prepare_() so
that the window is always a fixed aspect ratio and can be freely
resized.