Initial support for Switch shader (W/ BEARD BUG, READ COMMIT DESC)
* Added SampleShader.frag, SampleShader.vert derived from
SampleShader.glsl from the MiiShaderExample in the NintendoSDK.
  - Slightly modified, not completely original.
* Support this shader in Shader.cpp.

* Add ALLLLLL of the material values from nn::mii::Material and apply
them in Shader.cpp.
  - Could that be moved somewhere else...?

KNOWN FLAWS:

* FACE BEARDS WILL HAVE THE WRONG SKIN TONE ....
And potentially other things that are meant to go onto the faceline
texture.
This is because the Switch shader expects the faceline color's alpha
value to be 0. The only solution would be to apply it in FFL itself,
however since the FFL shader doesn't act the same if you do this, I just
added a hack to do it for most Miis in the shader.
* GLASSES DO NOT REFLECT LIKE THEY SHOULD I THINK...?
One problem with this is that the glass mesh does NOT have any normals
in FFLResMiddle/FFLResHigh, but it does on the Switch. They would have
to be added.
* UHHH THE BACK.. OF HAIR... UHHHHHH
Yyyyyeah the back of Jasmine's hair is dark in the MiiEdit applet and in
the VariableIconBody icons, however, in the IconBody icons it isn't. I
haven't been able to track down why it does there, since the shader
SEEMS the same, the uniforms are ALLLLL perfect, may come down to the
attributes... it may, in fact, be...
* i_Parameter is MISSING!
This is a vertex attribute that nn::mii provides, that this doesn't.
Apparently all it's used for is to be mixed into the specular factor for
the hair only... or, something. I can't easily find where it's being
passed in nn::mii, can't find its name at all.

continuing along with other changes:
* Add support for the shader to have access to the CharModel.

* Remove scale from Model class, not using it right now.
* Add ability for reading Mii data from a folder to be able to read
multiple Mii data types, all types supported by
pickupCharInfoFromRenderRequest.
* Modify pickupCharInfoFromRenderRequest to copy encoded studio data to
a temporary buffer and decode to that.
- Previously it'd decode it in place, which works but leaves the
buffer unusable for next time.