Update FFL and shader to improve mask accuracy, add blank expression
Hey gang, remember the "if(color.a == 0.0f) discard" line I added to the
FFLShader in order to improve the mask, in... some... way? Well that
broke mask alpha values, specifically it couldn't clear alpha with the
"fill" shape because it had an alpha of 0 and kept getting ignored...
great... Well with this FFL update and the FFLShader.frag update it
should be good. Should be.