Add xlu_depth_mask drawStageMode
This drawStageMode will "punch out" the depth from the DrawOpa stage
into the DrawXlu stage, and not draw the body.
In other words, if your Mii's eyes, glasses... are covered by its hair,
for example, that option will do DrawXlu with those areas punched out.

This is ideal for if you want to store Mii images; you can store the
DrawXlu result, then  the DrawXlu stage for each mask instead of all of
the composed images at once, then compose those two on the web browser
or something. Maybe. Not actually sure if that will be used.

Also remove unnecessary includes.