Update deps, move pickupCharInfoFromData out, fix a_color anomaly
* Solve this anomaly that's been kinda sorta not really happening with
the FFLDefaultShader and a_color.
- a_color is an attribute that sets rim lighting parameters for
non-aniso meshes (not hair).
- This attribute has some interesting behavior - sometimes it's fully
there and other times its stride is set to 0 but it has one value that's
meant to be constant and... I thiiinkkk OpenGL and GX2 account for this
behavior????? Not fully consistent though. Abood's code handled this
better but
- but NOW I've made rim width and everything else that used a_color,
use constant values from the fragment shader when you set a uniform to
one of two values, and that shoooulld account for all possibilities in
the FFL resource.
- I can't say this was necessary here, it will definitely be necessary
for FFL in bgfx/Vulkan/whatever but, here in this case it made me feel
better to account for this fringe case. Plus, this commit resolves one
very very slight inaccuracy with the hair that may or may not have even
been intentional and it's when using the beanie hair type so, that will
look correct now. I guess.

* Moved pickupCharInfoFromRenderRequest into DataUtils.cpp. Tiny part of
a refactor I hope I can begin here.

* Update FFL with texture callback and other unrelated changes, RIO and
ninTexUtils with MSVC fixes I think?