Preliminary light direction support, set NWF light direction (one to one matching!)
NWF renders from heeeere are... almost one to one matching after this.
The body uses a more normal light direction in NWF and I think the body
model position is a bit different but fflmakeicon ones are mostly fine
except for likeee this extra outline NWF has but yes. yes